+++ src/lib/payments/payWithFiat/getAuthorization.ts
@@ -1,0 +1,11 @@
+const getAuthorization = async (orderID: string) => {
+  const authorizationResponse = await fetch(`/api/onramp/authorization?orderId=${orderID}`)
+  if (!authorizationResponse) throw 'Error retrieving order'
+
+  const authorization = await authorizationResponse.json()
+  if (!authorization) throw 'Error interpreting authorization data'
+
+  return authorization
+}
+
+export default getAuthorization
